Extract ElephantListener into separate class
